Time: 3 ms (45.98%), Space: 8.4 MB (91.32%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -20,7 +20,7 @@ public:
 
         }
 
-        return num ==  (long)x;
+        return num == x;
 
         
     }
